# Use mol.GetBonds() instead of for loop