@KnathanM 2024-06-21
Yep, the result of message passing is equivalent.

My test:
```
import chemprop
import rdkit
import torch
mol = rdkit.Chem.MolFromSmiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)O)C") # Penicillin
dp = chemprop.data.MoleculeDatapoint(mol)
ds = chemprop.data.MoleculeDataset([dp])
dl = chemprop.data.build_dataloader(ds, batch_size=1)
batch = next(iter(dl))
mg, *_ = batch
checkpoint_path =  "tests/data/example_model_v2_regression_mol.ckpt"
mpnn = chemprop.models.MPNN.load_from_checkpoint(checkpoint_path)
mpnn.message_passing(mg, None)
```

On `main` it gives:
```
tensor([[0.0146, 0.0000, 0.0711,  ..., 0.1050, 0.1969, 0.0000],
        [0.0836, 0.0000, 0.0000,  ..., 0.2042, 0.3485, 0.2322],
        [0.0000, 0.0000, 0.1735,  ..., 0.1548, 0.2860, 0.0000],
        ...,
        [0.0000, 0.0398, 0.2541,  ..., 0.0511, 0.0206, 0.0000],
        [0.0000, 0.0000, 0.0603,  ..., 0.0074, 0.1134, 0.0000],
        [0.0146, 0.0000, 0.0711,  ..., 0.1050, 0.1969, 0.0000]],
       grad_fn=<ReluBackward0>)
```

On `faster-bond-iteration` it gives:
```
tensor([[0.0146, 0.0000, 0.0711,  ..., 0.1050, 0.1969, 0.0000],
        [0.0836, 0.0000, 0.0000,  ..., 0.2042, 0.3485, 0.2322],
        [0.0000, 0.0000, 0.1735,  ..., 0.1548, 0.2860, 0.0000],
        ...,
        [0.0000, 0.0398, 0.2541,  ..., 0.0511, 0.0206, 0.0000],
        [0.0000, 0.0000, 0.0603,  ..., 0.0074, 0.1134, 0.0000],
        [0.0146, 0.0000, 0.0711,  ..., 0.1050, 0.1969, 0.0000]],
       grad_fn=<ReluBackward0>)
```